case to uint64_t when convering gfid to inode, otherwise it can get garbled on 32 bit machines.
BUG: 2923
Reviewed-on: http://review.gluster.com/753
Tested-by: Gluster Build System <jenkins@build.gluster.com>

+++ xlators/storage/posix/src/posix-helpers.c
@@ -255,7 +255,7 @@ posix_fill_ino_from_gfid (xlator_t *this, struct iatt *buf)
                 goto out;
         }
         for (i = 15; i > (15 - 8); i--) {
-                temp_ino += buf->ia_gfid[i] << j;
+		temp_ino += (uint64_t)(buf->ia_gfid[i]) << j;
                 j += 8;
         }
         buf->ia_ino = temp_ino;
